requirements.txt
# Core data science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.0.0
scipy>=1.7.0

# Machine learning
lightgbm>=3.3.0
tensorflow>=2.8.0
xgboost>=1.5.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.5.0

# API and web
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0

# Utilities
joblib>=1.1.0
pyyaml>=6.0
category-encoders>=2.4.0
shap>=0.40.0
optuna>=2.10.0

# Development
jupyter>=1.0.0
ipykernel>=6.0.0
black>=22.0.0
flake8>=4.0.0

streamlit>=1.28.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.15.0
numpy>=1.24.0